test verifcation threshold

+++ contracts/default-registrar/src/msg.rs
@@ -1,4 +1,5 @@
 use cosmwasm_schema::cw_serde;
+use cosmwasm_std::Decimal;
 
 /// Message type for `instantiate` entry_point
 // TODO: change this to array
@@ -6,7 +7,7 @@ use cosmwasm_schema::cw_serde;
 pub struct InstantiateMsg {
     pub name_nft_addr: String,
     pub verifier_pubkeys: Vec<String>,
-    pub verification_threshold: u64,
+    pub verification_threshold: Decimal,
 }
 
 /// Message type for `execute` entry_point
@@ -25,7 +26,7 @@ pub enum ExecuteMsg {
         verifier_addr: String,
     },
     SetVerificationThreshold {
-        threshold: u64,
+        threshold: Decimal,
     },
 }
